cut down to key correlation functions
