+++ NAMESPACE
@@ -1,3 +1,8 @@
 # Generated by roxygen2: do not edit by hand
 
+export(cor2cor)
+export(fastCor)
 export(nearestPSDcor)
+import(RcppArmadillo)
+importFrom(Rcpp,sourceCpp)
+useDynLib(bigsimr, .registration = TRUE)
