+++ R/RcppExports.R
@@ -1,31 +1,15 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+CXX_nearPDcor <- function(G, tau = 1e-5, iter_outer = 200L, iter_inner = 20L, maxit = 200L, tol = 1e-2, err_tol = 1e-6, sigma1 = 1e-4) {
+    .Call(`_bigsimr_CXX_nearPDcor`, G, tau, iter_outer, iter_inner, maxit, tol, err_tol, sigma1)
+}
+
 cor2cor <- function(X, CASE) {
     .Call(`_bigsimr_cor2cor`, X, CASE)
 }
 
 hermite <- function(x, n) {
     .Call(`_bigsimr_hermite`, x, n)
-}
-
-nearPDcor <- function(G, tau = 1e-5, iter_outer = 200L, iter_inner = 20L, maxit = 200L, tol = 1e-2, err_tol = 1e-6, sigma1 = 1e-4) {
-    .Call(`_bigsimr_nearPDcor`, G, tau, iter_outer, iter_inner, maxit, tol, err_tol, sigma1)
-}
-
-rcor <- function(d, k = 1L) {
-    .Call(`_bigsimr_rcor`, d, k)
-}
-
-rmvn <- function(n, mu, sigma) {
-    .Call(`_bigsimr_rmvn`, n, mu, sigma)
-}
-
-rmvuu <- function(n, rho) {
-    .Call(`_bigsimr_rmvuu`, n, rho)
-}
-
-rmvu <- function(n, rho, min, max) {
-    .Call(`_bigsimr_rmvu`, n, rho, min, max)
 }
