+++ man/cor2cor.Rd
@@ -1,15 +1,9 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cor2cor.R, R/experimental.R
+% Please edit documentation in R/cor2cor.R
 \name{cor2cor}
 \alias{cor2cor}
 \title{Convert from one correlation to another}
 \usage{
-cor2cor(
-  rho,
-  from = c("pearson", "spearman", "kendall"),
-  to = c("pearson", "spearman", "kendall")
-)
-
 cor2cor(
   rho,
   from = c("pearson", "spearman", "kendall"),
@@ -24,7 +18,5 @@ cor2cor(
 \item{to}{Output correlation matrix}
 }
 \description{
-Convert from one correlation to another
-
 Convert from one correlation to another
 }
