+++ man/fastCor.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/fastCor.R
+\name{fastCor}
+\alias{fastCor}
+\title{Use the fastest methods available for computing each type of correlation}
+\usage{
+fastCor(x, y = NULL, method = c("pearson", "kendall", "spearman"))
+}
+\arguments{
+\item{x}{A matrix or vector}
+
+\item{y}{A vector}
+
+\item{method}{The type of correlation to compute}
+}
+\description{
+Use the fastest methods available for computing each type of correlation
+}
