--- man/hermite.Rd
@@ -1,16 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/hermite.R
-\name{hermite}
-\alias{hermite}
-\title{Computes the Hermite polynomial}
-\usage{
-hermite(x, n)
-}
-\arguments{
-\item{x}{a numeric vector giving the values at which the Hermite polynomial should be evaluated.}
-
-\item{n}{an integer giving the degree of the Hermite polynomials}
-}
-\description{
-Computes the Hermite polynomial
-}
